requirements.txt
attrdict==2.0.0
certifi==2018.10.15
chardet==3.0.4
confusable-homoglyphs==3.2.0
Django==2.1.11
django-bootstrap3==11.0.0
django-jquery==3.1.0
django-registration==3.0
djangorestframework==3.9.1
django-filter==2.2.0
markdown==3.1.1
idna==2.7
lru-dict==1.1.6
parsimonious==0.8.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pytz==2018.5
requests==2.20.0
rlp==1.0.3
six==1.11.0
toolz==0.9.0
urllib3==1.24.2
websockets==5.0.1